// src/enums/offer-status.enum.ts
export enum OfferStatus {
  PENDING = 'pending',
  ACCEPTED = 'accepted',
  REJECTED = 'rejected',
  WITHDRAWN = 'withdrawn',
  COUNTER_OFFERED = 'counter_offered',
  PENDING_COMPLETION_APPROVAL = 'pending_completion_approval',
  COMPLETION_ACCEPTED = 'completion_accepted',
  COMPLETION_REJECTED = 'completion_rejected'
}